backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
pymongo==4.0.1
Werkzeug==2.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0